Add error handling for user, product, and comment operations

+++ backend/src/db/queries.ts
@@ -21,16 +21,27 @@ export const getUserById = async (id: string) => {
 }
 
 export const updateUser = async (id: string, data: Partial<NewUser>) => {
+  const existingUser = await getUserById(id);
+  if(!existingUser) {
+    throw new Error (`User with id: ${id} not found`)
+  }
+
   const [user] = await db.update(users).set(data).where(eq(users.id, id)).returning();
   return user;
 }
 
 // upsert => create or update
 export const upsertUser = async (data: NewUser) => {
-  const existingUser = await getUserById(data.id);
-  if (existingUser) return updateUser(data.id, data);
+  const [user] = await db
+    .insert(users)
+    .values(data)
+    .onConflictDoUpdate({
+      target: users.id,
+      set: data,
+    })
+    .returning();
 
-  return createUser(data);
+  return user;
 }
 
 // product queries
@@ -40,10 +51,12 @@ export const createProduct = async (data: NewProduct) => {
   return product;
 }
 
-export const getAllProducts = async () => {
+export const getAllProducts = async (limit = 20, offset = 0) => {
   return db.query.products.findMany({
     with: { user: true },
-    orderBy: (products, {desc}) => [desc(products.createdAt)] // square brackets is required because drizzle orm's order by expects an array;
+    orderBy: (products, { desc }) => [desc(products.createdAt)],
+    limit,
+    offset,
   });
 };
 
@@ -60,22 +73,34 @@ export const getProductById = async (id: string) => {
   });
 };
 
-export const getProductsByUserId = async (userId: string) => {
+export const getProductsByUserId = async (userId: string, limit = 20, offset = 0) => {
   return db.query.products.findMany({
     where: eq(products.userId, userId),
     with: {
       user: true,
     },
     orderBy: (products, {desc}) => [desc(products.createdAt)],
+    limit,
+    offset,
   });
 };
 
 export const updateProduct = async (id: string, data: Partial<NewProduct>) => {
+  const existingProduct = await getProductById(id);
+  if (!existingProduct) {
+    throw new Error (`Product with id: ${id} not found`);
+  }
+
   const [product] = await db.update(products).set(data).where(eq(products.id, id)).returning();
   return product;
 };
 
 export const deleteProduct = async (id: string) => {
+  const existingProduct = await getProductById(id);
+  if (!existingProduct) {
+    throw new Error (`Product with id: ${id} not found`);
+  }
+
   const [product] = await db.delete(products).where(eq(products.id, id)).returning();
   return product;
 }
@@ -88,6 +113,11 @@ export const createComment = async (data: NewComment) => {
 }
 
 export const deleteComment = async (id: string) => {
+  const existingComment = await getCommentById(id);
+  if (!existingComment) {
+    throw new Error (`Product with id: ${id} not found`);
+  }
+
   const [comment] = await db.delete(comments).where(eq(comments.id, id)).returning();
   return comment;
 }
